+++ spec/javascripts/govuk_publishing_components/analytics/explicit-cross-domain-links.spec.js
@@ -28,7 +28,8 @@ describe('Explicit cross-domain linker', function () {
       decorate: function () {}
     }
 
-    explicitCrossDomainLinks = new GOVUK.Modules.ExplicitCrossDomainLinks()
+    element = $('<a href="#">')
+    explicitCrossDomainLinks = new GOVUK.Modules.ExplicitCrossDomainLinks(element[0])
 
     spyOn(window.gaplugins, 'Linker').and.returnValue(linker)
     spyOn(linker, 'decorate').and.callFake(function (url) {
@@ -43,13 +44,9 @@ describe('Explicit cross-domain linker', function () {
   })
 
   describe('when a cross-domain link is clicked', function () {
-    beforeEach(function () {
-      element = $('<a href="#">')
-    })
-
     it('modifies the link href to append cookie_consent parameter "not-engaged" if cookies_preferences_set cookie is "false"', function () {
       GOVUK.cookie('cookies_preferences_set', 'false')
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('href')).toEqual('#')
       window.GOVUK.triggerEvent(element[0], 'click')
       expect(element.attr('href')).toEqual('#?cookie_consent=not-engaged')
@@ -58,7 +55,7 @@ describe('Explicit cross-domain linker', function () {
 
     it('modifies the link href to append cookie_consent parameter "not-engaged" if cookies_preferences_set cookie is not set', function () {
       GOVUK.cookie('cookies_preferences_set', null)
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('href')).toEqual('#')
       window.GOVUK.triggerEvent(element[0], 'click')
       expect(element.attr('href')).toEqual('#?cookie_consent=not-engaged')
@@ -68,7 +65,7 @@ describe('Explicit cross-domain linker', function () {
     it('modifies the link href to append cookie_consent parameter "reject" if usage cookies have been rejected', function () {
       GOVUK.cookie('cookies_preferences_set', 'true')
       GOVUK.setConsentCookie({ usage: false })
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('href')).toEqual('#')
       window.GOVUK.triggerEvent(element[0], 'click')
       expect(element.attr('href')).toEqual('#?cookie_consent=reject')
@@ -81,7 +78,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.setConsentCookie({ usage: true })
         trackers = [{ ga_mock: 'foobar' }]
 
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('href')).toEqual('#')
         window.GOVUK.triggerEvent(element[0], 'click')
 
@@ -94,7 +91,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.setConsentCookie({ usage: true })
         trackers = []
 
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('href')).toEqual('#')
         window.GOVUK.triggerEvent(element[0], 'click')
         expect(element.attr('href')).toEqual('#?cookie_consent=accept')
@@ -105,7 +102,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.cookie('cookies_preferences_set', 'true')
         GOVUK.setConsentCookie({ usage: true })
         window.ga = undefined
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('href')).toEqual('#')
         window.GOVUK.triggerEvent(element[0], 'click')
         expect(element.attr('href')).toEqual('#?cookie_consent=accept')
@@ -120,11 +117,12 @@ describe('Explicit cross-domain linker', function () {
                '<input type="hidden" name="key" value="value" />' +
                '<button type="submit">Create a GOV.UK account</button>' +
              '</form>')
+      explicitCrossDomainLinks = new GOVUK.Modules.ExplicitCrossDomainLinks(element[0])
     })
 
     it('modifies the form action to append cookie_consent parameter "not-engaged" if cookies_preferences_set cookie is "false"', function () {
       GOVUK.cookie('cookies_preferences_set', 'false')
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('action')).toEqual('/somewhere')
       window.GOVUK.triggerEvent(element[0], 'submit')
 
@@ -133,7 +131,7 @@ describe('Explicit cross-domain linker', function () {
 
     it('modifies the form action to append cookie_consent parameter "not-engaged" if cookies_preferences_set cookie is not set', function () {
       GOVUK.cookie('cookies_preferences_set', null)
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('action')).toEqual('/somewhere')
       window.GOVUK.triggerEvent(element[0], 'submit')
       expect(element.attr('action')).toEqual('/somewhere?cookie_consent=not-engaged')
@@ -142,7 +140,7 @@ describe('Explicit cross-domain linker', function () {
     it('modifies the form action to append cookie_consent parameter "reject" if usage cookies have been rejected', function () {
       GOVUK.cookie('cookies_preferences_set', 'true')
       GOVUK.setConsentCookie({ usage: false })
-      explicitCrossDomainLinks.start(element)
+      explicitCrossDomainLinks.init()
       expect(element.attr('action')).toEqual('/somewhere')
       window.GOVUK.triggerEvent(element[0], 'submit')
       expect(element.attr('action')).toEqual('/somewhere?cookie_consent=reject')
@@ -153,7 +151,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.cookie('cookies_preferences_set', 'true')
         GOVUK.setConsentCookie({ usage: true })
         trackers = [{ ga_mock: 'foobar' }]
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('action')).toEqual('/somewhere')
         window.GOVUK.triggerEvent(element[0], 'submit')
         expect(element.attr('action')).toEqual('/somewhere?cookie_consent=accept&_ga=abc123')
@@ -163,7 +161,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.cookie('cookies_preferences_set', 'true')
         GOVUK.setConsentCookie({ usage: true })
         trackers = []
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('action')).toEqual('/somewhere')
         window.GOVUK.triggerEvent(element[0], 'submit')
         expect(element.attr('action')).toEqual('/somewhere?cookie_consent=accept')
@@ -173,7 +171,7 @@ describe('Explicit cross-domain linker', function () {
         GOVUK.cookie('cookies_preferences_set', 'true')
         GOVUK.setConsentCookie({ usage: true })
         window.ga = undefined
-        explicitCrossDomainLinks.start(element)
+        explicitCrossDomainLinks.init()
         expect(element.attr('action')).toEqual('/somewhere')
         window.GOVUK.triggerEvent(element[0], 'submit')
         expect(element.attr('action')).toEqual('/somewhere?cookie_consent=accept')