+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add wrapper to the component card to ensure link spacing ([PR #2753](https://github.com/alphagov/govuk_publishing_components/pull/2753))
 - Add new logic for counting links/sections on a second level browse page on page view #2733 ([PR #2733](https://github.com/alphagov/govuk_publishing_components/pull/2733))
 * Update cross domain linking script to use init ([PR #2747](https://github.com/alphagov/govuk_publishing_components/pull/2747))
+* Remove DWP email from attachment accessible format request pilot([PR #2764](https://github.com/alphagov/govuk_publishing_components/pull/2764))
 
 ## 29.7.0